Feature/pos printing (#2)
* feat(print): domain layer for POS ticket printing (BSP-7)

Pure-domain vertical slice for printing the receipt ("cuenta") from a bill ID:
- dto.Ticket / TicketItem / TicketBusinessInfo
- dto.PrintTicketRequest { OpenBillID, Copies }
- ports.ReceiptPrinter (takes *dto.Ticket, never bytes)
- domain errors ErrOpenBillNotFound / ErrTicketPrintFailed
- PrintService.PrintTicket: load open bill, map to Ticket (backend-computed
  totals), print N copies
- generated MockReceiptPrinter + unit tests (success, not-found, printer error,
  copies, mapping/totals)


* feat(print): hand-rolled ESC/POS renderer + golden tests (BSP-8)

Pure-Go renderer (no external ESC/POS lib) that encodes a dto.Ticket to bytes,
with all control sequences contained in one file for later swap:
- init, codepage select (ESC t), center/left, double-size header, bold TOTAL,
  feed + full/partial/none cut
- rune-aware column formatter (left name / right money) at configurable width
  (48 ≈ 80mm, 32 ≈ 58mm)
- COP money formatting ($, '.' thousands) and word-wrap
- UTF-8 -> codepage transcoding (CP850/CP858/CP1252) for accents/ñ via x/text,
  unsupported runes replaced
- golden tests (80mm/58mm) + control-sequence, cut-mode, wrap, padLR, format,
  nil-ticket assertions


* feat(print): device transport + ReceiptPrinter adapter + config (BSP-9)

Adapter that renders a ticket and ships the bytes over a swappable transport:
- Transport interface + NewTransport factory (file | tcp | windows | serial)
- fileTransport (ticket-NNNN.escpos, macOS dev) and tcpTransport (RAW :9100)
- transport_windows.go (//go:build windows) RAW-spooler stub for Phase B, and
  transport_windows_stub.go (//go:build !windows) so macOS keeps compiling;
  verified GOOS=windows builds the device package
- ReceiptPrinter implements ports.ReceiptPrinter (render -> mutex-serialized
  transport write), plus OptionsFromConfig and NewReceiptPrinterFromConfig
- config: PRINTER_TRANSPORT/TARGET/WIDTH_MM/CODEPAGE/CUT (default file/CP850/
  partial/80mm) and BUSINESS_*/TICKET_* identity, all optional
- device unit tests: render+write, transport error, file sequencing, factory
  selection, options mapping


* feat(print): device print endpoint + edge wiring + API docs (BSP-10)

- DeviceHandler.PrintTicketHandler binds {open_bill_id, copies} and maps domain
  errors to the contract: 200 {printed:true}, 404 bill_not_found,
  409 printer_unavailable, 400 validation_error
- cmd/main.go edge branch: build the printer from PRINTER_* config, construct
  PrintService with TicketBusinessInfo, register
  POST /api/device/print (JWT + orders:read). Edge mode only; cloud never
  registers it. If printer init fails, log and skip the route so the node still
  boots (frontend falls back to browser print)
- docs/api/device.md + curl examples


* fix: limiting the number of copies for the ticket

* fix: linter errors fixed

---------

+++ internal/domain/service/invoice_submission_service.go
@@ -170,7 +170,7 @@ func invoiceBackoff(attempts int) time.Duration {
 	if attempts > 16 {
 		return invoiceBackoffMax
 	}
-	d := invoiceBackoffBase << uint(attempts)
+	d := invoiceBackoffBase << attempts
 	if d > invoiceBackoffMax {
 		return invoiceBackoffMax
 	}
